+++ auto_route/lib/auto_route.dart
@@ -1,7 +1,6 @@
 export 'src/auto_route_wrapper.dart';
 export 'src/extended_navigator.dart';
 export 'src/route_guard.dart';
-export 'src/route_matcher.dart';
 export 'src/router_utils.dart';
 export 'src/transitions_builders.dart';
-export 'src/auto_router.dart';
+export 'src/route_def.dart';