+++ auto_route/lib/src/route_def.dart
@@ -1,0 +1,27 @@
+
+import 'package:auto_route/auto_route.dart';
+
+class RouteDef {
+  final String template;
+  final List<Type> guards;
+  final RouterBuilder innerRouter;
+  final Pattern pattern;
+  final Type page;
+
+  RouteDef(
+      this.template, {
+        this.page,
+        this.guards,
+        this.innerRouter,
+      }) : pattern = _buildPathPattern(template);
+
+  static Pattern _buildPathPattern(String template) {
+    return '^${template.replaceAllMapped(RegExp(r':([^/]+)|([*])'), (m) {
+      if (m[1] != null) {
+        return '(?<${m[1]}>[^/]+)';
+      } else {
+        return ".*";
+      }
+    })}';
+  }
+}