+++ auto_route/lib/src/router_base.dart
@@ -3,9 +3,10 @@ part of 'extended_navigator.dart';
 typedef AutoRouteFactory = Route<dynamic> Function(RouteData data);
 typedef RouterBuilder<T extends RouterBase> = T Function();
 
-class RouterBase {
-  RouterBase({this.routes});
-  final List<RouteDef> routes;
+abstract class RouterBase {
+  List<RouteDef> get routes;
+
+  Map<Type, AutoRouteFactory> get pagesMap;
 
   Set<String> get allRoutes => routes.map((e) => e.template).toSet();
 
@@ -14,20 +15,19 @@ class RouterBase {
     assert(settings != null);
     var match = findFullMatch(settings);
     if (match != null) {
-      var matchResult = match.copyWith(name: "${basePath ??= ''}${settings.name}") as MatchResult;
+      var namePrefix = basePath?.isNotEmpty == true ? "$basePath" : '';
+      var matchResult = match.copyWith(name: "$namePrefix${settings.name}") as RouteMatch;
       RouteData data;
       if (matchResult.isParent) {
         data = _ParentRouteData(
           matchResult: matchResult,
-          initialRoute: matchResult.rest.toString(),
-          router: matchResult.routeDef.router(),
+          initialRoute: matchResult.restAsString,
+          router: matchResult.routeDef.innerRouter(),
         );
       } else {
         data = RouteData(matchResult);
       }
-      print("pushing: $data");
-
-      return matchResult.routeDef.builder(data);
+      return pagesMap[matchResult.routeDef.page](data);
     }
     return null;
   }
@@ -37,26 +37,24 @@ class RouterBase {
   // Router().onGenerateRoute becomes Router()
   Route<dynamic> call(RouteSettings settings) => onGenerateRoute(settings);
 
-  MatchResult findFullMatch(RouteSettings settings) {
+  RouteMatch findFullMatch(RouteSettings settings) {
     // deep links are  pre-matched
-    MatchResult matchResult;
-    if (settings is MatchResult) {
-      matchResult = settings;
+    if (settings is RouteMatch) {
+      return settings;
     } else {
-      final matcher = RouteMatcher(settings);
+      var matcher = RouteMatcher(settings);
       for (var route in routes) {
         var match = matcher.match(route, fullMatch: true);
         if (match != null) {
-          matchResult = match;
-          break;
+          return  match;
         }
       }
     }
-    return matchResult;
+    return null;
   }
 
-  List<MatchResult> allMatches(RouteMatcher matcher) {
-    var matches = <MatchResult>[];
+  List<RouteMatch> allMatches(RouteMatcher matcher) {
+    var matches = <RouteMatch>[];
     for (var route in routes) {
       var matchResult = matcher.match(route);
       if (matchResult != null) {
@@ -68,39 +66,6 @@ class RouterBase {
     }
     return matches;
   }
-
-  MatchResult match(RouteSettings settings) {
-    final matcher = RouteMatcher(settings);
-    for (var route in routes) {
-      var match = matcher.match(route);
-      if (match != null) {
-        return match;
-      }
-    }
-    return null;
-  }
 }
 
-class RouteDef {
-  final String template;
-  final List<Type> guards;
-  final RouterBuilder router;
-  final AutoRouteFactory builder;
-  final Pattern pattern;
 
-  RouteDef(this.template, {this.guards, this.router, this.builder}) : pattern = _buildPathPattern(template);
-
-  static Pattern _buildPathPattern(String template) {
-    return '^${template.replaceAllMapped(RegExp(r':([^/]+)|([*])'), (m) {
-      if (m[1] != null) {
-        return '(?<${m[1]}>[^/]+)';
-      } else {
-        return ".*";
-      }
-    })}';
-  }
-
-//  bool get hasGuards => guards?.isNotEmpty == true;
-//
-//  bool get isParent => router != null;
-}