+++ example/lib/router/router.dart
@@ -1,11 +1,8 @@
 import 'package:auto_route/auto_route_annotations.dart';
-import 'package:example/generic_a.dart';
-import 'package:example/generic_b.dart';
 import 'package:example/screens/home_screen.dart';
-import 'package:example/screens/login_screen.dart';
 import 'package:example/screens/unknown_route.dart';
 import 'package:example/screens/users/sub/profile_screen.dart';
-import 'package:example/screens/users/sub/user_details.dart';
+import 'package:example/screens/users/sub/user_posts.dart';
 import 'package:example/screens/users/users_screen.dart';
 
 export 'package:auto_route/auto_route.dart';
@@ -14,17 +11,17 @@ export 'router.gr.dart';
 
 @MaterialAutoRouter(
   routes: <AutoRoute>[
-    CustomRoute(page: HomeScreen, initial: true),
-    MaterialRoute<GenericA<GenericB<int>>>(
+    MaterialRoute(page: HomeScreen, initial: true),
+    MaterialRoute(
       path: '/users/:id',
       page: UsersScreen,
-      children: [
-        CustomRoute(path: '/', page: UserDetails),
-        MaterialRoute(path: '/profile', page: ProfileScreen),
+      children: <AutoRoute>[
+        // path: '/' is the same as setting initial to true
+        MaterialRoute(path: '/', page: ProfileScreen),
+        MaterialRoute(path: '/posts', page: PostsScreen),
       ],
     ),
-    CustomRoute<bool>(path: "/login", page: LoginScreen, fullscreenDialog: true),
-    MaterialRoute(path: "*", page: UnknownRouteScreen)
+    MaterialRoute(path: '*',page: UnknownRouteScreen)
   ],
 )
 class $Router {}