+++ example/lib/screens/home_screen.dart
@@ -3,8 +3,6 @@ import 'package:example/router/router.dart';
 import 'package:flutter/material.dart';
 
 class HomeScreen extends StatelessWidget {
-//  final  TypedFunction onSelected;
-//  HomeScreen(this.onSelected);
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -17,21 +15,7 @@ class HomeScreen extends StatelessWidget {
               child: FlatButton(
                 child: Text("Users Screen"),
                 onPressed: () async {
-                  ExtendedNavigator.of(context).pushNamed("/users/1?foo=bar");
-//                  AutoRouter.of(context).push(Routes.initalRoute);
-//                ExtendedNavigator.of(context).pushNamed(Routes.secondScreen);
-//                  ExtendedNavigator.ofRouter<Router>().pushNamed(
-//                    Routes.secondScreen,
-//                    arguments: SecondScreenArguments(message: 'title'),
-//                    onReject: (guard) {
-//                      showDialog(
-//                        context: context,
-//                        builder: (_) => AlertDialog(
-//                          title: Text('You need to be logged in'),
-//                        ),
-//                      );
-//                    },
-//                  );
+                  ExtendedNavigator.root.pushNamed('/users/1?filter=active');
                 },
               ),
             ),