+++ example/lib/screens/login_screen.dart
@@ -2,10 +2,6 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
 class LoginScreen extends StatelessWidget {
-  final double id;
-
-  LoginScreen({this.id = 20.0});
-
   @override
   Widget build(BuildContext context) {
     return WillPopScope(
@@ -20,8 +16,6 @@ class LoginScreen extends StatelessWidget {
             child: Text("Login"),
             onPressed: () async {
               ExtendedNavigator.of(context).pop<bool>(true);
-              // .pushReplacementNamed(Routes.homeScreen);
-              // ExtendedNavigator.rootNavigator.pop(true);
             },
           ),
         ),