+++ example/lib/screens/unknown_route.dart
@@ -5,6 +5,7 @@ class UnknownRouteScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: AppBar(),
       backgroundColor: Colors.grey,
       body: Center(
         child: Text(