+++ example/lib/screens/users/users_screen.dart
@@ -1,31 +1,23 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:auto_route/auto_route_annotations.dart';
 import 'package:flutter/material.dart';
 
 class UsersScreen extends StatelessWidget {
-  final String id;
-  final int score;
-
   const UsersScreen({
-    this.id,
-    this.score,
-    double limit = 0.0,
+    @PathParam() this.id,
+    @QueryParam('filter') this.filterFromQuery = "none",
   });
+
+  final int id;
+  final String filterFromQuery;
 
   @override
   Widget build(BuildContext context) {
-    // var route = ModalRoute.of(context);
-    // var settings = route.settings as ExtendedRouteSettings;
-//    print(RouteData.of(context));
-
     return Scaffold(
-        appBar: AppBar(
-          title: Text("Users Details $id score:$score"),
-        ),
-//      body: FlatButton(
-//          child: Text("push deep link"),
-//          onPressed: () async {
-//            ExtendedNavigator.of(context).pushNamed(Routes.usersScreen);
-//          }),
-        body: NestedNavigator());
+      appBar: AppBar(title: Text("Users id: $id, filter: $filterFromQuery")),
+      // this navigator will obtain it's router
+      // on it's own
+      body: NestedNavigator(name: 'usersRouter'),
+    );
   }
 }